pfeng: Detect HIF reset posibility for Slave
When Slave starts, it shall analyze the state of HIF channel HW.
The state of channel HW affects the channel init procedure.
There can be detected the following states:
1) clean: no channel registers were used before
2) ungraceful: the channel registers already contains correct
   setup
3) invalid: some of the channel registers can contains incorrect
   values

The state 1) doesn't require anything, the driver can continue
on it's start-up without any magic.

The state 2) requires a special processing, to cover use-case,
when the driver or OS was ungracefully stopped.

The state 3) signals some inconsistency in channel setup.

Issue: AAVB-7397
Signed-off-by: Jan Petrous <jan.petrous@nxp.com>