pfeng: Fix WB ring dump formatting by using GET accessors, cleanup
The WB ring dump is wrongly using the SET instead of the GET accessors
to print the 'seqnum' and 'buflen' content, fix that.

Fix also the argument name of the GET accessors to prevent such
confusion in the future, and add some new lines to make the print
routine calls more readable.

Issue: AAVB-8078
Signed-off-by: Claudiu Manoil <claudiu.manoil@nxp.com>