pfeng: Fix iteration of invalid WB BDs in pfe_hif_ring_find_wb_entry()
To find the first invalid BD, the function needs to iterate
again the ring starting from the position of the fist valid
BD + 1. Without this fix, some Rx BD iteration cases would
fail, e.g. when the fist valid BD is on the last position in
the ring.

Reported-by: Yibo Liu <yibo.liu@nxp.com>
Issue: AAVB-8636
Signed-off-by: Claudiu Manoil <claudiu.manoil@nxp.com>