pfeng: Fix wrong error return of ungraceful reset when the Master cleans up the BD rings state
When the while loop is skipped in pfe_hif_chnl_rx_to_head(),
or in pfe_hif_chnl_tx_to_head(), because the Master did the
job of cleaning up the state of the BD rings, 'ret' is no longer
touched by the code logic and so both functions wrongly return the
error code they take as default.

To fix this, 'ret' should no longer default to error.

Reported-by: Yibo Liu <yibo.liu@nxp.com>
Issue: AAVB-8636
Signed-off-by: Claudiu Manoil <claudiu.manoil@nxp.com>