pfeng: Remove unused accessors in hal.h
Removed the `hal_ip_ready_set` and `hal_ip_ready_get` accessors
in hal.h, used for communication between PFE Master and Slave
instances through the GENCTRL3 register. The access to the
register is now done through syscon, as for the other GPRs,
or via NVMEM if enabled.

Issue: AAVB-4948
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>